Support multiple of 8 image widths
Integer division in Rust rounds down, so for example a 192px wide image
would be marked as having 25 pixels by the old code. This caused a
crash, so now I check if there are any extraneous bytes for a row.

If a row is a non-integer multiple of 8, it's padded. The resulting
value should mean there's no overflow in this image drawing code.

I should probably refactor this.