Transform Command into a fixed array and a length and call send only once
The previous way of calling send_commands() directly caused a lot of
code duplication and calls for the SPI interface layer. To work around
this issue I'm now assembling the data into a fixed sized array (Rust
does not allow varying return types on match) and a length which then is
unconditionally and only once passed over to the interface for
transmission. It's the most idiomatic way I could come up with without
grabbing for the nightly or unsafe toolset.

Signed-off-by: Daniel Egger <daniel@eggers-club.de>