New, more consistent terminal font
This font is taken from the ATIx_550_8x8.png font image from the
romfonts repo. A bitmap can be found here:
https://github.com/spacerace/romfont/blob/master/font-images/ATIx550_8x8.png

I don't have hardware to test with tat the moment so I might have messed
up the character mappings. I'd appreciate if someone could test on a
physical display and see how this looks.

I'm not including any of the special characters defined in that font. If
we want to add them, I'd suggest we do that in a future PR.

Closes #135